refactor(webui): split control + runs handlers
handlers_control.go (821 -> 381 LOC) by extracting fork/rewind
handlers, adapter/model meta endpoints, run-launcher methods,
loadPipelineYAML and option-projection helpers into dedicated
files (handlers_fork.go, handlers_meta.go, server_launcher.go,
server_helpers.go).

handlers_runs.go (1352 -> 319 LOC) by extracting projection
helpers (buildStepDetails, computeGanttPositions, runToSummary,
eventToSummary, deduplicate/enrich helpers) into
handlers_runs_projection.go, run detail page rendering with its
output card / linked URL / template var helpers into
handlers_run_detail.go, CSV/JSON export into
handlers_runs_export.go, and JSON write helpers into
json_helpers.go. The running-runs collection branch in
handleRunsPage was hoisted into Server.collectRunningRuns.

Partial of #1498. Server struct split (D) and dropping the
internal/pipeline import from internal/webui (B) deferred to a
follow-up: gate types, fork manager, and gate registry still
require pipeline types in DTOs and handler bodies, so a clean
break needs a new narrow-interface layer rather than a file
move.
